requirements.txt
Flask==0.10.1
flask-cors==3.0.2
PyJWT==1.4.0
Flask-SQLAlchemy==1.0
requests==2.7.0
